requirements.txt
streamlit>=1.27.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.9.0
scikit-learn>=1.1.0
pillow>=9.2.0
mistralai>=0.0.7
tensorflow>=2.10.0  # Optional: Only needed for GAN/VAE functionality